pyproject.toml
[project]
name = "modelx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.11.12",
    "dagshub>=0.6.3",
    "fastapi>=0.121.2",
    "flask>=3.1.2",
    "geopy>=2.4.1",
    "ipykernel>=7.1.0",
    "mlflow>=3.6.0",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyaml>=25.7.0",
    "pymongo[srv]>=4.15.4",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.1",
]

[tool.uv.workspace]
members = [
    "venv",
]
